+++ src/db/repositories/reviews-repository.ts
@@ -5,28 +5,77 @@ import type {
 } from '@/domain/entities/review'
 import { db } from '@/db'
 import { schema } from '@/db/schema'
-import { and, eq } from 'drizzle-orm'
+import type { GetReviewsServiceInput } from '@/domain/services/reviews/get-reviews'
+import { and, desc, eq, getTableColumns } from 'drizzle-orm'
+import type { UpdateReviewInput } from '@/domain/services/reviews/update-review'
+import type { GetDetailedReviewsInput } from '@/domain/services/reviews/get-detailed-reviews'
 
 export async function insertReview(params: InsertReviewModel) {
   return db.insert(schema.reviews).values(params).returning()
 }
 
-export async function insertReviewReply(params: InsertReviewReplyModel) {
-  return db.insert(schema.reviewReplies).values(params).returning()
-}
-
-export async function getReviewById(id: string) {
-  return db.select().from(schema.reviews).where(eq(schema.reviews.id, id))
-}
-
-export async function deleteReviewReply(params: DeleteReviewReplyModel) {
+export async function selectReviewsWithUser({
+  mediaType,
+  tmdbId,
+}: GetReviewsServiceInput) {
   return db
-    .delete(schema.reviewReplies)
+    .select({
+      ...getTableColumns(schema.reviews),
+      user: {
+        id: schema.users.id,
+        username: schema.users.username,
+        imagePath: schema.users.imagePath,
+      },
+    })
+    .from(schema.reviews)
     .where(
       and(
-        eq(schema.reviewReplies.id, params.id),
-        eq(schema.reviewReplies.userId, params.userId)
+        eq(schema.reviews.tmdbId, tmdbId),
+        eq(schema.reviews.mediaType, mediaType)
       )
     )
+    .leftJoin(schema.users, eq(schema.reviews.userId, schema.users.id))
+}
+
+export async function selectReviews({
+  userId,
+  language,
+  limit,
+}: GetDetailedReviewsInput) {
+  return db
+    .select({
+      ...getTableColumns(schema.reviews),
+      user: {
+        id: schema.users.id,
+        username: schema.users.username,
+        imagePath: schema.users.imagePath,
+      },
+    })
+    .from(schema.reviews)
+    .orderBy(desc(schema.reviews.createdAt))
+    .limit(Number(limit))
+    .where(
+      and(
+        userId ? eq(schema.reviews.userId, userId) : undefined,
+        language ? eq(schema.reviews.language, language) : undefined
+      )
+    )
+    .leftJoin(schema.users, eq(schema.reviews.userId, schema.users.id))
+}
+
+export async function deleteReview(id: string) {
+  return db.delete(schema.reviews).where(eq(schema.reviews.id, id)).returning()
+}
+
+export async function updateReview({
+  id,
+  rating,
+  review,
+  hasSpoilers,
+}: UpdateReviewInput) {
+  return db
+    .update(schema.reviews)
+    .set({ rating, review, hasSpoilers })
+    .where(eq(schema.reviews.id, id))
     .returning()
 }