+++ src/domain/services/user-items/get-user-items.ts
@@ -1,0 +1,15 @@
+import { selectUserItems } from '@/db/repositories/user-item-repository'
+
+export type GetUserItemsInput = {
+  userId: string
+  status: 'WATCHED' | 'WATCHLIST' | 'WATCHING'
+}
+
+export async function getUserItemsService({
+  userId,
+  status,
+}: GetUserItemsInput) {
+  const userItems = await selectUserItems({ userId, status })
+
+  return { userItems }
+}