+++ src/domain/services/users/update-user-subscription.spec.ts
@@ -1,12 +1,11 @@
 import { describe, expect, it } from 'vitest'
 import { makeUser } from '@/test/factories/make-user'
 
-import { updateUserImageService } from './update-user-image'
 import { faker } from '@faker-js/faker'
 import { UserNotFoundError } from '@/domain/errors/user-not-found'
 import { updateUserSubscriptionService } from './update-user-subscription'
 
-describe('update username subscription', () => {
+describe('update user subscription', () => {
   it('should be able to update user subscription type', async () => {
     const user = await makeUser()