+++ src/http/routes/reviews.ts
@@ -1,13 +1,25 @@
 import type { FastifyInstance } from 'fastify'
 import type { ZodTypeProvider } from 'fastify-type-provider-zod'
 
-import { createReviewController, createReviewReplyController } from '../controllers/review-controller'
 import {
-  createReviewRequestSchema,
+  createReviewController,
+  deleteReviewController,
+  getDetailedReviewsController,
+  getReviewsController,
+  updateReviewController,
+} from '../controllers/review-controller'
+import {
+  createReviewBodySchema,
   createReviewResponseSchema,
-  createReviewReplyRequestSchema,
-  createReviewReplyResponseSchema,
+  getDetailedReviewsQuerySchema,
+  getDetailedReviewsResponseSchema,
+  getReviewsQuerySchema,
+  getReviewsResponseSchema,
+  reviewParamsSchema,
+  updateReviewBodySchema,
+  updateReviewResponse,
 } from '../schemas/reviews'
+import { verifyJwt } from '../middlewares/verify-jwt'
 
 export async function reviewsRoute(app: FastifyInstance) {
   const reviewsTag = 'Reviews'
@@ -15,28 +27,77 @@ export async function reviewsRoute(app: FastifyInstance) {
   app.after(() =>
     app.withTypeProvider<ZodTypeProvider>().route({
       method: 'POST',
-      url: '/reviews/create',
+      url: '/review',
+      onRequest: [verifyJwt],
       schema: {
         description: 'Create a review',
         tags: [reviewsTag],
-        body: createReviewRequestSchema,
+        body: createReviewBodySchema,
         response: createReviewResponseSchema,
+        security: [
+          {
+            bearerAuth: [],
+          },
+        ],
       },
       handler: createReviewController,
     })
   )
-  
+
   app.after(() =>
     app.withTypeProvider<ZodTypeProvider>().route({
-      method: 'POST', 
-      url: '/reviews/reply',
+      method: 'GET',
+      url: '/reviews',
       schema: {
-        description: 'Create a review reply',
+        description: 'Get reviews',
         tags: [reviewsTag],
-        body: createReviewReplyRequestSchema,
-        response: createReviewReplyResponseSchema,
+        querystring: getReviewsQuerySchema,
+        response: getReviewsResponseSchema,
       },
-      handler: createReviewReplyController,
+      handler: getReviewsController,
+    })
+  )
+
+  app.after(() =>
+    app.withTypeProvider<ZodTypeProvider>().route({
+      method: 'DELETE',
+      url: '/review/by/:id',
+      schema: {
+        description: 'Delete review by id',
+        tags: [reviewsTag],
+        params: reviewParamsSchema,
+      },
+      handler: deleteReviewController,
+    })
+  )
+
+  app.after(() =>
+    app.withTypeProvider<ZodTypeProvider>().route({
+      method: 'PUT',
+      url: '/review/by/:id',
+      schema: {
+        description: 'Update review by id',
+        tags: [reviewsTag],
+        params: reviewParamsSchema,
+        body: updateReviewBodySchema,
+        response: updateReviewResponse,
+      },
+      handler: updateReviewController,
+    })
+  )
+
+  app.after(() =>
+    app.withTypeProvider<ZodTypeProvider>().route({
+      method: 'GET',
+      url: '/detailed/reviews',
+      schema: {
+        description: 'Get detailed reviews',
+        tags: [reviewsTag],
+        query: getDetailedReviewsQuerySchema,
+        response: getDetailedReviewsResponseSchema,
+      },
+      handler: (request, reply) =>
+        getDetailedReviewsController(request, reply, app.redis),
     })
   )
 }